Move cow tests to their own file
